JustPod: Consolidate functions
This commit changes to use typeglobs for various functions that behave
identically to other ones, so that future changes can be made in one
place.

It also causes several functions to call a base function with an
argument that distinguishes this one from similar ones.  Again, future
changes need be done only in the base function.