BlackBox.pm: Preserve special case for =item * foo bar
It is legal to say

=item * This is text on the same line as an asterisk

It is not documented in perlpodspec, but that asterisk is specially
handled to turn into a bullet.  Russ Allberry in <87y3gtcwa2.fsf@hope.eyrie.org>
indicates this has been the case for a very long time.  So we are
preserving the behavior.

But this requires special handling in JustPod, and this commit stops the
deletion of the hash element that indicates special handling is needed,
so that JustPod can know about it.