RTF.pm: Improve detection of perl identifiers
This code is looking for potential perl symbols to declare off-limits
for spell checking.  But it was doing a really poor job of non-ascii
ones.  I converted to use the best available Unicode properties that
deal with identifiers, and if the perl this is running on is too early,
just the typical ASCII letters and digits.