Fix strip_verbatim_indent and HTML
This commit is sort of a guess to resolve github issue #85.  It simply
comments out a line that caused indentation to not be stripped, that
otherwise would have been as specified by the strip_verbatim_indent()
method.

The line, if executed, would ignore the stripping, apparently in favor
of any VerbatimFormatted 'code' furnished.  This 'code' appears to be an
XML tag that can be overridden by the caller of the routines.  It does
get overridden in the HTML.pm module to be the standard HTML
<pre>...</pre> sequence, thereby causing any strip_verbatim_indent()
call to be ignored.

The logic behind the line may have been: "They're using a custom way of
specifying verbatim formatting, so we should ignore anything else that
affects it".  But, there's no warning that specifying this overrides the
other, and it's not incompatible that the user would want to strip some
indentation before using the custom verbatim formatting.

But there may have some other reason instead, which does not get
exercised by the test suite.  Hence the tentativeness of this commit
message.